+++ app/controllers/submission-controller.js
@@ -10,6 +10,7 @@ var request = require( 'request' );
 var express = require( 'express' );
 var router = express.Router();
 var routerUtils = require( '../lib/router-utils' );
+var url = require( 'url' );
 // var debug = require( 'debug' )( 'submission-controller' );
 
 module.exports = function( app ) {
@@ -61,8 +62,6 @@ function submit( req, res, next ) {
         .then( function( survey ) {
             submissionUrl = communicator.getSubmissionUrl( survey.openRosaServer ) + query;
             credentials = userModel.getCredentials( req );
-
-            // first check if authentication is required and if so get the Basic or Digest Authorization header
             return communicator.getAuthHeader( submissionUrl, credentials );
         } )
         .then( function( authHeader ) {