+++ app/lib/communicator/communicator.js
@@ -145,6 +145,11 @@ function getAuthHeader( url, credentials ) {
     };
 
     return new Promise( function( resolve ) {
+        // Don't bother making Head request first if token was provided.
+        if ( credentials.bearer ) {
+            resolve( 'Bearer ' + credentials.bearer );
+        }
+        // Check if Basic or Digest Authorization header is required and return header if so.
         var req = request( options, function( error, response ) {
             if ( !error && response && response.statusCode === 401 && credentials && credentials.user && credentials.pass ) {
                 // Using request's internal library we create an appropiate authorization header.
@@ -194,7 +199,7 @@ function getUpdatedRequestOptions( options ) {
     // set Authorization header
     if ( !options.auth ) {
         delete options.auth;
-    } else {
+    } else if ( !options.auth.bearer ) {
         // check first is DIGEST or BASIC is required
         options.auth.sendImmediately = false;
     }